requirements.txt
pandas==1.1.3
influxdb==5.3.1
pytz==2020.1
numpy==1.19.2
binance==0.3
matplotlib==3.3.2
python-decouple==3.4
python_dateutil==2.8.1